Fix bugs and clean-up
The focus of this change is to iron out the issues during the flash
test. Here are the issues addressed:

1./ Wrong fields when parsing efuse_boot_header configure because
of 'strncmp' matching. Parse the configuration lines into token and
value pointer with spaces trimmed, then use 'strcmp'.

2./ Issues with accessing device port. Use the default block mode.

3./ CRC missing in erase packet.

4./ Wrong packet length in the flash data packet. Should include
the length of the targeted address.

5./ SHA256 packet has incomplete values.

Other than that, the following improvements:

a./ Rewrite hand-shake code with dynamic memory to hold the number
of '0x55' in the stream.

b./ Move redudant debug prints under #ifdef DEBUG.

c./ Expose dump_hex so that other file can use it.

d./ replace '\t' with spaces.